==> putin_2021/03_Putin_2021.ipynb <==
# загружаю нужные библиотеки

import pandas as pd

work_putin = pd.read_excel('C:/00_Projects/2021/Putin_2021/putin.xlsx', sheet_name='data')

(work_putin
 .groupby(['region', 'federal_district'], as_index=False)
 .agg({'street': 'count'})
 .rename(columns={'street': 'street_number'})
 .sort_values(['federal_district', 'street_number'], ascending=[True, False])
)

#.to_excel('C:/00_Projects/2021/Putin_2021/temp.xlsx')

(work_putin[(
    (work_putin.sub_city == 'city') &
    (work_putin.city != 'Грозный')
)]
 .groupby('federal_district', as_index=False)
 .agg({'city_population': 'mean'})
 #.rename(columns={'street': 'street_number'})
 .sort_values('city_population', ascending=False)
)

work_putin.sample()

(work_putin
 .groupby('year', as_index=False)
 .agg({'street': 'count'})
 #.rename(columns={'street': 'street_number'})
 .sort_values('year', ascending=False)
).to_excel('C:/00_Data/temp.xlsx', index=False)

(work_putin[(
    (work_putin.sub_city == 'city') &
    (work_putin.city != 'Грозный')
)]
 .groupby('federal_district', as_index=False)
 .agg({'city_population': 'mean'})
 #.rename(columns={'street': 'street_number'})
 .sort_values('city_population', ascending=False)
)
